added query for users

+++ src/users/users.service.ts
@@ -1,37 +1,62 @@
-import { Injectable } from '@nestjs/common';
+import { UserRepository } from './user.repository';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/user.dto';
 import { User } from './user.entity';
-import { DeleteResult } from 'typeorm/query-builder/result/DeleteResult';
-import { UpdateResult } from 'typeorm/query-builder/result/UpdateResult';
+import { GetUsersFilterDto } from './dto/get-users-filter.dto';
 
 @Injectable()
 export class UsersService {
   constructor(
     @InjectRepository(User)
     private readonly usersRepository: Repository<User>,
-  ) {
+    @InjectRepository(UserRepository)
+    private userRepository: UserRepository,
+  ) { 
   }
 
   public async save(createUserDto: CreateUserDto): Promise<User> {
     return this.usersRepository.save(createUserDto);
   }
 
-  public async update(id: number, createUserDto: CreateUserDto): Promise<UpdateResult> {
-    return this.usersRepository.update({id}, createUserDto);
+  public async update(id: number, createUserDto: CreateUserDto): Promise<User> {
+    const user = await this.usersRepository.findOne({
+      where: { id }
+    });
+    if (!user) {
+      throw new HttpException('User with this ID not found', HttpStatus.NOT_FOUND);
+    }
+    this.usersRepository.update({id}, createUserDto);
+    const updatedUser = await this.usersRepository.findOne({
+      where: { id }
+    });
+    return updatedUser;
   }
 
-  public async find(): Promise<User[]> {
-    return this.usersRepository.find();
+  async getUsers(filterDto: GetUsersFilterDto): Promise<User[]> {
+    return this.userRepository.getUsers(filterDto);
   }
 
   public async findOne(id: number): Promise<User> {
-    return this.usersRepository.findOne(id);
+    const user = await this.usersRepository.findOne({
+      where: { id }
+    });
+    if (!user) {
+      throw new HttpException('Chat with this ID not found', HttpStatus.NOT_FOUND);
+    }
+    return user;
   }
 
-  public async delete(id: number): Promise<DeleteResult> {
-    return this.usersRepository.delete(id);
+  public async delete(id: number): Promise<User> {
+    const user = await this.usersRepository.findOne({
+      where: { id }
+    });
+    if (!user) {
+      throw new HttpException('Chat with this ID not found', HttpStatus.NOT_FOUND);
+    }
+    await this.usersRepository.delete(id);
+    return user;
   }
 
   public async getUserByEmail(email: string): Promise<User> {
